==> src/Integrated/04_CellTypeInference.Rmd <==
---
title: "Celltype Inference"
author: "Karsten Bach"
date: '`r Sys.Date()`'
output:
  html_notebook:
    toc: true
    toc_depth: 3
    toc_float: true
    theme: united
    highlight: tango
    code_folding: hide
---
***

```{r, message=FALSE,warning=FALSE}
library(scran)
library(scater)
library(plyr)
library(dplyr)
library(knitr)
library(ggplot2)
library(cowplot)
library(Matrix)
library(irlba)
library(schex)
library(ggplot2)
source("../functions.R")

sce <- readRDS("../../data/Integrated/Robjects/SCE_combined.rds")

rD <- rowData(readRDS("../../data/Tumorigenesis/Robjects/SCE.rds"))
rmgenes <- rownames(rD)[!rD$KeepForHvg]
rmgenes <- rownames(sce)[rownames(sce) %in% rmgenes]
```

```{r, fig.width=9, fig.height=9, warning=FALSE, message=FALSE}
plotUMAP(sce, colour_by="Condition", text_by="Cluster")
```

# Major Groups

First cells where grouped into four broad categories:

- Epithelial (EpCam+)
- Immune (Ptprc+)
- Fibroblasts (Col3a1)
- Stroma (Various markers)

## Epithelial
```{r, fig.width=9, fig.height=9, warning=FALSE, message=FALSE}
plotUMAP(sce, colour_by="Epcam", text_by="Cluster")
epithelial <- c(11,30,5, # Basal
		19,21,40,6,9, #Luminal
		1) # Tumor
epiCluster <- paste0("C",epithelial)
```

## Immune
- Most cells marked by Ptprc (Cd45)
- Neutrophils marked by S100a8
- PlasmaCells Jchain
- B-Cells Cd19 
```{r, fig.width=12, fig.height=4, warning=FALSE, message=FALSE}
immune <- c(15,16,27,29,3,37,39,8, # Lymphoid
	    23,24,25,26,32,34,36,38,41,43,45) # Myeloid

plotExpression(sce, x="Cluster", features=c("Ptprc"))
plotExpression(sce, x="Cluster", features=c("S100a8"))
neutros <- c(4) # not ptprc pos but S100a8
plotExpression(sce, x="Cluster", features=c("Jchain"))
plasma <- c(35) # not ptprc but Jchain
immuneCluster <- paste0("C",c(immune,neutros,plasma))
```

# Fibroblasts

- Mostly marked by Col1a1/Col3a1
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
plotUMAP(sce, colour_by="Col1a1", text_by="Cluster")
fibroblasts <- c(17,18,2,22,31,33,42,44,46)
fibroblastsCluster <- paste0("C",fibroblasts)
```


# Stroma
- Bit more heterogeneous not a single marker
- C14 Schwann Cell (Mbp, Mpz)
- All endothelial cells below also have Cdh5
- C10, 12, 28, 7 cells (Emcn)
- C7 lymphatic endothelium (Nts, Lyve1)
- C13, 20 are pericytes (Notch3, Acta2)
```{r, fig.width=12, fig.height=4, warning=FALSE, message=FALSE}
# Stroma
# 14 is Schwann Cell (Mbp, Mpz)
# 41 lymphatic endothelium (Nts, # not really Lyve1 but Nts marked the lymphatic endothelium in prev)
# 10,12,28,7 are endothelium (Cdh5)
# 7 Higd1b,Rgs5,Gm13889, Rgs4 no idea, will label as stroma
# C13 and 20 are possibly pericytes
plotExpression(sce, x="Cluster", features=c("Mbp"))
plotExpression(sce, x="Cluster", features=c("Cdh5"))
plotExpression(sce, x="Cluster", features=c("Nts"))
plotExpression(sce, x="Cluster", features=c("Esm1"))
plotExpression(sce, x="Cluster", features=c("Emcn"))
plotExpression(sce, x="Cluster", features=c("Notch3"))
plotExpression(sce, x="Cluster", features=c("Acta2"))
stroma <- c(14, # Schwann
	    10,12,28, # Endothelial
	    7, #Lymphatic Endothelium
	    13,20) # Pericyte

stromaCluster <- paste0("C",stroma)
sce$MajorGroups <- mapvalues(as.character(sce$Cluster),
		       from=c(epiCluster, immuneCluster, fibroblastsCluster, stromaCluster),
		       to=c(rep("Epithelial",length(epiCluster)),
			    rep("Immune",length(immuneCluster)),
			    rep("Fibroblast",length(fibroblastsCluster)),
			    rep("Stroma",length(stromaCluster))))
```

This then results in the following annotation of the larger groups

```{r, fig.width=9, fig.height=9, warning=FALSE, message=FALSE}
plotUMAP(sce, colour_by="MajorGroups", text_by="Cluster")
```

# Finer Annotation

- Some of the groups can be resolved in smaller sub-compartments
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
tumorCells <- c(1)
luminal <- c(19,21,40,6,9)
basal <- c(11,30,5)

lymphoid <- c(15,16,27,29,3,37,39,8,35)
myeloid <- c(23,24,25,26,32,34,36,38,41,43,45,4) 

luminal <- paste0("C",luminal)
basal <- paste0("C",basal)
myeloid <- paste0("C",myeloid)
lymphoid <- paste0("C",lymphoid)
tumorCells <- paste0("C",tumorCells)
# 
sce$Groups <- mapvalues(as.character(sce$Cluster),
			from=c(lymphoid, myeloid, basal, luminal, tumorCells, fibroblastsCluster, stromaCluster),
			to=c(rep("Lymphoid",length(lymphoid)),
			     rep("Myeloid",length(myeloid)),
			     rep("Basal",length(basal)),
			     rep("Luminal",length(luminal)),
			     rep("TumorEpithel",length(tumorCells)),
			     rep("Fibroblast",length(fibroblastsCluster)),
			     rep("Stroma",length(stromaCluster))))
plotUMAP(sce, colour_by="Groups", text_by="Cluster")
```

# Cell Type mapping
- Cell types were inferred based on marker gene expresssion and the previous annotation.
- Most cell types are quite obvious.
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
# mark <- findMarkers(sce,sce$Cluster,pval.type="all",block=sce$Batch)

celltypes <- c(1:46)
celltypes <- paste0("C",celltypes)

#pDC because of Siglec-Hhttps://www.ncbi.nlm.nih.gov/pmc/articles/PMC3977341/
names(celltypes) <- c("Tm","Fb1","ILCs","Neutrophils","Bsl1",#5
		      "Hs","Ec1","CyclingT","Avp","Ec2",#10

		      "Bsl2","Ec3","Pericytes1","Swc","TcCTL",#15
		      "NK1","Fb2","Fb3","Hsp","Pericytes2",#20

		      "Hs","Fb4","Mo1","Mp1Dm","Tam1",#25
		      "Tam2","BCells","Ec4","TcNaive","BslG",#30

		      "Fb5","cDC2","Fb6","Mo2","PlasmaCells",#35
		      "migDC","CD81","DcsMbe","CD41","Lp",#40

		      "MastCells","Fb7","Mo3","Fb8","Mcs",#45
		      "Fb9")

sce$CellTypes <- mapvalues(sce$Cluster, from=celltypes, to=names(celltypes))
sort(table(sce$CellTypes), decreasing=TRUE)
plotUMAP(sce, colour_by="Groups", text_by="CellTypes")
plotUMAP(sce, colour_by="Groups", text_by="Cluster")
```

# QC Check
- As this will be the final annotation, I will one last time check that none of the clusters is funny in terms of QC criteria
- This shouldn't be a problem anymore as this has been done independently for the two datasets
```{r, fig.width=10, fig.height=3, warning=FALSE, message=FALSE}
plotColData(sce, x="CellTypes",y="UmiSums") + scale_y_log10() + theme(axis.text.x=element_text(size=9, angle=45, hjust=1, vjust=1))
plotColData(sce, x="CellTypes",y="GenesDetected") + scale_y_log10() + theme(axis.text.x=element_text(size=9, angle=45, hjust=1, vjust=1))
plotColData(sce, x="CellTypes",y="prcntMito") + theme(axis.text.x=element_text(size=9, angle=45, hjust=1, vjust=1))
plotColData(sce, x="CellTypes",y="DbltScore") + theme(axis.text.x=element_text(size=9, angle=45, hjust=1, vjust=1))
```

## Further Resolution of some clusters

### TcCTL
- This cluster contains a mix of NK cells and CD8 cells.
- Subclustering resolves this cluster into a CD8 cluster, one NK cell cluster (that I will most likely merge with the other NK cells) and 
  a cluster of cytotoxic lymphocytes, where the GEX profile is ambiguous and it is hard to label them definitely as CD8 T or NK
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="TcCTL")) +
    geom_point()
```

- Sub Clustering with SNN and louvain
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes=="TcCTL"]
m.cor <- reducedDim(sce.sub,"corrected")

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
igr <- buildSNNGraph(t(m.cor))
subc <- igraph::cluster_louvain(igr)
sce.sub$SubCluster <- factor(subc$membership)
plotUMAP(sce.sub,text_by="SubCluster",colour_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up",
			subset.row=!rownames(sce.sub)%in% rmgenes)

plotExpression(sce.sub, features=c("Klrb1c","Eomes","Klra8","Klra4","Gzma",
			       "Cd4","Cd8a","Cd3d"),x="SubCluster") +
    theme(axis.text.x=element_text(angle=45,hjust=1))
plot_grid(plotUMAP(sce.sub,colour_by="Klrb1c",text_by="SubCluster"),
plotUMAP(sce.sub,colour_by="Cd8a",text_by="SubCluster"),
plotUMAP(sce.sub,colour_by="Cd3d",text_by="SubCluster"),
plotUMAP(sce.sub,colour_by="Cd4",text_by="SubCluster")
)
# Based on the GEX and the fact that there is little DE between the sub clusters, I label them as follows
sce$CellTypes[sce$barcode %in% sce.sub$barcode] <- paste0(sce.sub$CellTypes,sce.sub$SubCluster)
sce$CellTypes[sce$CellTypes %in% c("TcCTL2","TcCTL4","TcCTL5")] <- "CD83"
sce$CellTypes[sce$CellTypes %in% c("TcCTL3","TcCTL6")] <- "CTLs"
sce$CellTypes[sce$CellTypes %in% c("TcCTL1")] <- "NK2"
```


### Bsl1 contains small group of doublets
- Contains some obvious structure, judging from the plot it looks like it clusters with some doublets
- The other subclusters are not kept as they don't have more than 10 DE genes

```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="Bsl1")) +
    geom_point()
```

- Sub Clustering with SNN and louvain
- 10 expresses Cd3d and Epcam
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes=="Bsl1"]
m.cor <- reducedDim(sce.sub,"corrected")

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
igr <- buildSNNGraph(t(m.cor))
subc <- igraph::cluster_louvain(igr)
sce.sub$SubCluster <- factor(subc$membership)
plotUMAP(sce.sub,text_by="SubCluster",colour_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up",
			subset.row=!rownames(sce.sub)%in% rmgenes)
plotExpression(sce.sub,x="SubCluster","Cd3d")
rmCells <- sce.sub$barcode[sce.sub$SubCluster=="10"]
```

### Fb1 contains a small group of myeloid cells
- Contains some weird grouping with a myeloid cluster
- The myeloid cluster are most likely poor quality cells from various groups / and / or doublets as they express markers for various myeloid cells at low levels
- This can also be seen if visualised as a UMAP with the myeloid cells where the randomly distribute across all diferent groups (not shown)
- The other Fb cluster are kept as Fb1 as they don't have enough DE genes
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="Fb1")) +
    geom_point()
```

```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes=="Fb1"]
m.cor <- reducedDim(sce.sub,"corrected")

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
igr <- buildSNNGraph(t(m.cor))
subc <- igraph::cluster_louvain(igr)
sce.sub$SubCluster <- factor(subc$membership)
plotUMAP(sce.sub,text_by="SubCluster",colour_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up",
			subset.row=!rownames(sce.sub)%in% rmgenes)
```
- Subcluster 12 expresses markers of macrophages, basophils, dendritic cells 
```{r, fig.width=12, fig.height=6, warning=FALSE, message=FALSE}
plotExpression(sce.sub, x="SubCluster",features=c("Lyz2","Adgre1","Ccr7","Il6","Arg1","Chil3"),ncol=1)
rmCells <- c(rmCells, sce.sub$barcode[sce.sub$SubCluster=="12"])
```

### Lp
- Contains some obvious structure, judging from the plot it looks like it clusters with some doublets
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="Lp")) +
    geom_point()
```

- Sub Clustering with SNN and louvain
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes=="Lp"]
m.cor <- reducedDim(sce.sub,"corrected")

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
igr <- buildSNNGraph(t(m.cor))
subc <- igraph::cluster_walktrap(igr,steps=20)
sce.sub$SubCluster <- factor(subc$membership)
p1 <- plotUMAP(sce.sub,colour_by="SubCluster",text_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up",
			subset.row=!rownames(sce.sub)%in% rmgenes)
mark.sig <- lapply(mark.sub,function(x) data.frame(x)[x$FDR<0.01,])
```

-4, 10 is Immune/LP doublet
-17 are Immune cells will be kept seperately
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
#plotExpression(sce.sub,x="SubCluster","Csn2")
#plotUMAP(sce.sub,colour_by="Csn2",text_by="SubCluster")
```

-7, 13,15 and 16 are Fb doublet
- 12 is Fb but also some Krts etc.
```{r, fig.width=10, fig.height=10, warning=FALSE, message=FALSE}
features <- c("Csn2","Aldh1a3","Krt5","Col3a1","Cd74","Cited1","Krt8","Krt17")
p0 <- plotExpression(sce.sub,x="SubCluster",features)
p0
```

- 1 and 11 are damaged cells, have either very low UMI or very low mitochondria count

```{r, fig.width=10, fig.height=7, warning=FALSE, message=FALSE}
plotColData(sce.sub,x="SubCluster",y="prcntMito")
plotColData(sce.sub,x="SubCluster",y="DbltScore")
```


- I remove 1,8 and 9 and the rest is kept as possible sub Lp clusters, but might me merged below
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
stopifnot(identical(sce.sub$barcode, sce$barcode[sce$barcode %in% sce.sub$barcode]))
sce$CellTypes[sce$barcode %in% sce.sub$barcode] <- paste0(sce.sub$CellTypes,sce.sub$SubCluster)
rmCells <- c(rmCells,sce.sub$barcode[sce.sub$SubCluster %in% c("1","11","4","12","10",
							       "7","13",
							       "15","16")])
```

### Mcs
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="Mcs")) +
    geom_point()
```

- Sub Clustering with SNN and louvain
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes=="Mcs"]
m.cor <- reducedDim(sce.sub,"corrected")

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
igr <- buildSNNGraph(t(m.cor))
subc <- igraph::cluster_walktrap(igr,steps=30)
sce.sub$SubCluster <- factor(subc$membership)
plotUMAP(sce.sub,text_by="SubCluster",colour_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up",
			subset.row=!rownames(sce.sub)%in% rmgenes)
```

-6 express Ear2
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Ear2")
```

-3 are pDCs (Siglech)
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Siglech")
```

- I concatenate with Mcs and redefine 6 and 3
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
stopifnot(identical(sce.sub$barcode, sce$barcode[sce$barcode %in% sce.sub$barcode]))
sce$CellTypes[sce$barcode %in% sce.sub$barcode] <- paste0(sce.sub$CellTypes,sce.sub$SubCluster)
sce$CellTypes[sce$CellTypes=="Mcs3"] <- "pDC"
sce$CellTypes[sce$CellTypes=="Mcs6"] <- "MdC3"
```

- Clean Up, removing the doublets and damaged cells
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce <- sce[,!sce$barcode %in% rmCells]
```

### TcNaive
- This cluster contains a mix of Tregs, CD4 cells and CD8 T Cells
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
ggplot(data.frame(colData(sce)), aes(x=UMAP1, y=UMAP2, color=CellTypes=="TcNaive")) +
    geom_point()
```

- Sub Clustering with SNN and louvain
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes=="TcNaive"]
m.cor <- reducedDim(sce.sub,"corrected")

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
igr <- buildSNNGraph(t(m.cor))
subc <- igraph::cluster_louvain(igr)
sce.sub$SubCluster <- factor(subc$membership)
plotUMAP(sce.sub,text_by="SubCluster",colour_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up",
			subset.row=!rownames(sce.sub)%in% rmgenes)
```

- 1 are Tregs
- Ctla4, Foxp3, Tnfrsf18
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
head(data.frame(mark.sub[["1"]]),10)
plotExpression(sce.sub,x="SubCluster","Ctla4")
plotExpression(sce.sub,x="SubCluster","Tnfrsf18")
```

- 2,4-6 are grouped based on Cd8
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Cd8a")
```

- 3,7 are grouped based on Cd4
```{r, fig.width=12, fig.height=3, warning=FALSE, message=FALSE}
plotExpression(sce.sub,x="SubCluster","Cd4")
```


- I concatenate with Tcs and redefine 1 and 2
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
stopifnot(identical(sce.sub$barcode, sce$barcode[sce$barcode %in% sce.sub$barcode]))
sce$CellTypes[sce$barcode %in% sce.sub$barcode] <- paste0(sce.sub$CellTypes,sce.sub$SubCluster)
sce$CellTypes[sce$CellTypes=="TcNaive1"] <- "Tregs"
sce$CellTypes[sce$CellTypes %in% c("TcNaive2","TcNaive4","TcNaive5","TcNaive6")] <- "CD82"
sce$CellTypes[sce$CellTypes %in% c("TcNaive3","TcNaive7")] <- "CD42"
```

# Fb 5
- This cluster contains one obvious "outlier group" that represent fibroblasts from the tumour (CAFs)
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes=="Fb5"]
m.cor <- reducedDim(sce.sub,"corrected")

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
igr <- buildSNNGraph(t(m.cor))
subc <- igraph::cluster_louvain(igr)
sce.sub$SubCluster <- factor(subc$membership)
plotUMAP(sce.sub,text_by="SubCluster",colour_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up",
			subset.row=!rownames(sce.sub)%in% rmgenes,
			lfc=1)
lapply(mark.sub, function(mrks) rownames(mrks)[mrks$FDR<0.01])
```

- Fb52 is a cluster of tumor associated fibroblasts
- The rest I keep together as I wouldn't be able to interpret them
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
stopifnot(identical(sce.sub$barcode, sce$barcode[sce$barcode %in% sce.sub$barcode]))
sce$CellTypes[sce$barcode %in% sce.sub$barcode] <- paste0(sce.sub$CellTypes,sce.sub$SubCluster)
sce$CellTypes[sce$CellTypes=="Fb52"] <- "Cafs"
sce$CellTypes[sce$CellTypes %in% c("Fb51","Fb53","Fb54","Fb55","Fb56","Fb57","Fb58","Fb59")] <- "Fb5"
```


# Merging
- I finally run mergeCluster again to only maintain clusters with min 10 DE genes unless the clusters represent meaningful subgroups
  which is generally the case for some of the T Cell clusters
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
m <- logcounts(sce)
merged <- mergeCluster(m, factor(sce$CellTypes), removeGenes=rmgenes, block=sce$Batch,
		       maxRep = 30)
```

- This suggests merging some of the Lps in seperate groups
- I am keeping the T-Cell clusters seperatley as I know that they correspond to different states of T-Cells (unsure about CD81 and CD82, but will leave it like this for now)
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce$CellTypesFinal <- as.character(merged$NewCluster)
sce$CellTypesFinal[sce$CellTypesFinal=="Lp5.Lp6.Lp8.Lp9"] <- "Lp"
sce$CellTypesFinal[sce$CellTypesFinal=="Avp.Lp3.Lp14"] <- "Avd"
sce$CellTypesFinal[sce$CellTypesFinal=="Lp2"] <- "LpBsl"
sce$CellTypesFinal[sce$CellTypesFinal=="Lp17.PlasmaCells"] <- "PlasmaCells"
sce$CellTypesFinal[sce$CellTypesFinal=="Mcs4"] <- "MdC2"
sce$CellTypesFinal[sce$CellTypesFinal=="Mcs1.Mcs2"] <- "cDC1"
sce$CellTypesFinal[sce$CellTypesFinal=="DcsMbe.Mcs5"] <- "MdC1"
sce$CellTypesFinal[sce$CellTypesFinal=="CD41.CD42.CD82.Tregs"] <- sce$CellTypes[sce$CellTypes%in% c("CD41","CD42","CD82","Tregs")]
sce$CellTypesFinal[sce$CellTypesFinal=="CD83.CTLs.NK2"] <- sce$CellTypes[sce$CellTypes%in% c("CTLs","NK2","CD83")]
sce$CellTypesFinal[sce$CellTypesFinal=="Mp1Dm"] <- "Mo3" # These are merged as the only difference seems to be sequencing depth
sce$MajorGroups[sce$CellTypesFinal=="PlasmaCells"] <- "Immune"
sce$Groups[sce$CellTypesFinal=="PlasmaCells"] <- "Lymphoid"

```

## NK clusters
- The two NK clusters only show 3 genes DE, I will merge them for simplicity
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
sce.sub <- sce[,sce$CellTypes %in% c("NK1","NK2")]
m.cor <- reducedDim(sce.sub,"corrected")

ump.cor <- umap::umap(m.cor, random_state=42)
reducedDim(sce.sub,type="UMAP") <- ump.cor$layout[,1:2]
set.seed(42)
sce.sub$SubCluster <- sce.sub$CellTypes
plotUMAP(sce.sub,text_by="SubCluster",colour_by="SubCluster")
mark.sub <- findMarkers(sce.sub, sce.sub$SubCluster,
			block=sce.sub$Batch,
			pval.type="all",
			direction="up",
			subset.row=!rownames(sce.sub)%in% rmgenes,
			lfc=1)
head(data.frame(mark.sub[[1]]),4)

sce$CellTypesFinal[sce$CellTypesFinal %in% c("NK1","NK2")] <- "NK"
```

# Final Annotation and Save
```{r, fig.width=7, fig.height=7, warning=FALSE, message=FALSE}
saveRDS(sce,file="../../data/Integrated/Robjects/SCE_combined_celltypes.rds")
sessionInfo()
```
